Revert "Remove all branching around OS_TYPE"
This reverts commit 59a9c29c782aec9ccb25724b5d4f4c93dd29e7f7.

# Conflicts:
#	stemcell_builder/lib/prelude_agent.bash
#	stemcell_builder/lib/prelude_apply.bash
#	stemcell_builder/stages/bosh_azure_chrony/apply.sh
#	stemcell_builder/stages/bosh_ntp/apply.sh
#	stemcell_builder/stages/static_libraries_config/apply.sh

Merge conflict notes:
- 248de3db, f3202115, stemcell_builder/stages/bosh_azure_chrony/apply.sh: Should this file's code be Ubuntu-only or not?

+++ stemcell_builder/stages/system_kernel_modules/apply.sh
@@ -28,4 +28,6 @@ options nouveau modeset=0
 alias nouveau off
 alias lbm-nouveau off' >> $chroot/etc/modprobe.d/blacklist-nouveau.conf
 
-rm -rf $chroot/lib/modules/*/kernel/zfs $chroot/usr/src/linux-headers-*/zfs
+if [[ $OS_TYPE == 'ubuntu' ]]; then
+  rm -rf $chroot/lib/modules/*/kernel/zfs $chroot/usr/src/linux-headers-*/zfs
+fi
